Externalized Bootstrap Dependency
For the sake of importing different CSS and JavaScript libraries into
the project, it makes much more sense to rely on NPM to store the
dependencies rather then bloating the git repository.

To keep the size of the Docker image to a minimum, I've taken advantage
of Docker's multi-stage building process to install the dependencies and
potentially build the rest of the package, should there ever be a need
to do so.